requirements.txt
pathway
groq
sentence-transformers
pyyaml
pandas
numpy
tqdm